Move KubectlApply to helpers package
